refactor(cli): move outputs rendering into separate component
