+++ packages/cdktf-cli/bin/cmds/ui/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./plan-element";
 export * from "./deploying-element";
+export * from "./output";
