+++ packages/cdktf-cli/bin/cmds/ui/deploy.tsx
@@ -3,12 +3,8 @@ import React, { Fragment, useState } from "react";
 import { Text, Box } from "ink";
 import Spinner from "ink-spinner";
 import ConfirmInput from "@skorfmann/ink-confirm-input";
-import { DeployingElement } from "./components";
-import {
-  DeployingResource,
-  TerraformOutput,
-  PlannedResourceAction,
-} from "./models/terraform";
+import { DeployingElement, Output } from "./components";
+import { DeployingResource, PlannedResourceAction } from "./models/terraform";
 import { Status, useTerraformState, useRunDeploy } from "./terraform-context";
 import { Plan } from "./diff";
 
@@ -45,33 +41,6 @@ export const DeploySummary = ({
         </Box>
       ))}
     </>
-  );
-};
-
-interface OutputConfig {
-  output: { [key: string]: TerraformOutput };
-}
-function sanitize(value: any) {
-  if (typeof value === "object") {
-    return JSON.stringify(value, null, 2);
-  }
-
-  return value;
-}
-export const Output = ({ output }: OutputConfig): React.ReactElement => {
-  return (
-    <Box flexDirection="column">
-      {Object.keys(output).map((key) => (
-        <Box key={key}>
-          <Text>
-            {key} ={" "}
-            {output[key].sensitive
-              ? "<sensitive>"
-              : sanitize(output[key].value)}
-          </Text>
-        </Box>
-      ))}
-    </Box>
   );
 };
